Rename general/mm.?pp -> general/mem_manager.?pp
