+++ general/mem_manager.hpp
